Clean up entries controller ...
Add date value and column when viewing a date collection.

Ensure a title column is there. It goes at the start if it's not.

Changed makeListableColumns() to just columns(). We probably wont ever need to get non-listable columns.

Columns are keyed so we can use ->has() etc on the collection. Keys are stripped in the controller.