Live preview for creating entries, and related things ...
Requesting the entry URL directly worked okay for existing entries, but can't work for new ones, they just 404. We'll just get the response from the entry itself.

We'll swap the new and old request before and after getting the entry's response.

The kernel no longer handles the whole request, which means we lose things like terminating handling, but that might be okay. It can return later if needed.

Also, the amp request was being set as part of the controller, which no longer gets hit now, so the logic for determining an amp route is moved.