Structure can have collections ...
A structure can also have "a collection". That would be the collection that uses this structure as "its" structure.

The collections specified on a structure are what ends up showing the page selector.

If a structure has "a collection", then the collections cannot be specified manually and is just an array (technically an illuminate\support\collection, yikes) with that one collection.