Simplify store removals, let the complicated ones do their own thing ...
Handle localizations separately from their parent classes.

When inserting a global, its localizable classes get their paths saved.
Likewise for structure and its trees.

When removing a localizable global, it knows to remove it from the set and update it.
Likewise when removing a tree. It removes it from the structure and updates it. It also removes the uris and routes.