Term titles get tracked ...
When a term is thrown into a markdown file on the fly, whatever value is there is considered the title. It gets normalized (slugified) and stored using that as the key.

If there are multiple entries with the same "normalized" value, the most recent one will be used as the title.

Users should maintain this consistency themselves. Either make the values consistent when editing files directly, or create an actual term with a title.