Taxonomies get automatically routed ...
Route, template, and layout are no longer set on the taxonomy.

Visiting /{collection}/{taxonomy}/{slug?} will load the taxonomy or a term.
Customizing taxonomy routes is now considered an advanced feature, so users should make their own routes. This is helpful to the majority, though.

The collection of terms will be sent to the template using the taxonomy's handle as the variable name.