Fix bard tests, and add enabled attribute ...
There was an error because the fieldtype checks whether it's a content field. It reads that from the field. The tests hadn't been updated to add a field.

The enabled attribute was being added to values, because replicator does it. Bard wants it in attributes, though.

The legacy data check should immediately fail if its a string, since that's a v3 thing.