DataRepository tweaks
Added tests for the method_exists that was added in 6a553d8

Mock method_exists with a technique from a stack overflow post. Hacky but neat.

Also make a contract and bind the repos in the container, since the tests were running the hardcoded ones in addition to the mocked ones.