LabeledValue object ...
Allows a fieldtype (eg. select) to return this object that will output the value when used as a string in Antlers. When used as an array you can get access to the value or label.

Also when a Value's augmented value is Arrayable, the parser can get values through there.