Don't store items in memory, or they'd be the same objects ...
eg. If you set a property on an entry, dont save it, and then do Entry::find(), that property would already be there.

This was added in a364f47 because we wanted to compare objects. We can (and should) just compare IDs (or later add an is() method like Eloquent models) rather than strict-comparing objects.