:sparkles: Better automatic autofocus
The `autofocus` attribute appeared to be conflicting with the HTML attribute, setting it on the component container instead of passing it through. I also added a check on the Text fieldtype so that if the fieldname is "title" it'll turn on autofocus. This should be what we want all the time.