Tree tweaks ...
As part of (the collection structure's) tree validation, only entries from that site should get automatically appended. In order to do that, we need to query against the site.
This causes an infinite loop when it happens during the stache being loaded.

(In the validation, we query entries by site, it tries to populate the index, it needs to get the structure, it makes the structure from disk, it validates itself, and there's the infinite loop.)

To get around this, the validation is moved to the getter, and cached with Blink to get around the potentially performance hit.

When saving a collection, add or remove the appropriate trees depending on what sites are chosen.

When a collection structure requests a site's tree that it doesn't have in its tree list yet, just make it, assuming its allowed by the collection. Only navs need to explicitly exist. Collection structures should just work depending on what sites the collection has.

Entry files should be ignored if they're in a site subdirectory corresponding to one that the collection hasn't asked for.