Link to the nav's showUrl ...
rather than the selected site's tree.

It's possible that the nav doesn't exist for the selected site, which would cause an error.
We could filter it out of the listing, but then you just wouldn't see it.

Better to just show all the navs. When you visit the showUrl, it will redirect appropriately.