Remove the fix ...
It was causing problems when saving. Messing up the collection references within the collection's structure. Since the stache was empty, sometimes where it needed to find the collection it would pull a fresh one, which would contain the existing structure defined in the file and things would be confused.

The multisite issue is still there, but is less of a priority. It only breaks for a moment.