Only query params with a colon (eg. is:somefield) ...
The loop here would call queryCondition on each parameter, but something would only happen if there was a colon. The third argument would fall back to false, and then the switch statement would never match anything.

But since d6b1d21, the $this->getQueryConditionValue() was added, which gets called on every iteration, even if it's not a parameter intended for querying.

This bug presented itself when doing {{ entries }} ... {{ /entries }} on a taxonomy term page. It translated to {{ query :builder="builder" }}, and the query builder instance was the value that caused an exception in getQueryConditionValue.