Use a separate exception handler for API ...
Swap the exception handler in a middleware, since in a service provider we don't have access to the request and it would be too early to conditionally rebind.

Since it's in a middleware, it'll only use that exception handler (and thus handle the json 404 responses) when it actually hits a real route. Added a wildcard 404 route to the API like we have in the CP.