The actual used multisite will be detected correctly
This PR will patch the following issue:
https://github.com/statamic/cms/issues/1874

The `absoluteUrl()` method form the Site class, will remove the `/` from the absolute url. 
This will lead to the behavior, that a string like
`mypage.com/datenschutz` 
will be detected as a `da` handle from a multisite. What it should have detected:
`mypage.com/da/`

When checking if the url starts with a handle from a multisite, we need to make sure that we add the `/` to the right, like done in the added `absoluteUrl()` methods. 

With this patch, it will be checked against:
`mypage.com/da/` instead of `mypage.com/daANYTHING`