Catch the DirectoryNotFoundException when clearing static cache
In the following cases the cache folder does not exist:

* A vanilla installation when the user just changed the caching strategy
to 'full' in `static_caching.php` and did not visit the website yet
* After the user flushed the cache, since the command deletes the cache
folder

Just catch the exception and do nothing else with it. Printing an error
notice to the user is IMO not necessary because:

* Its normal behavior that the folder is not there as it got deleted by
the command itself.
* The cache is empty and this is exactly what the user expects from the
command

Closes #1972