Add error tag for the default laravel error bag
There are some cases, where forms won't be created via the Statamic form tag. If this case appears, there is no way in antlers to fetch those errors. 

If creating a form without the form tag, you can easily fetch errors like:
```
{{ error:name }} // Will return the message

{{ if {error:name }} Do something, add a class or similar {{ /if }}
```

**Would a PR like this be merged? Is there an interest in integrating this behavior into antlers?**

This is what I build for a client and would be a great starting point, but is clearly a draft at this state. Before finishing it off, I wanted to get some opinions first 😎