Improve overriding repositories ...
Users can now call Statamic::repository() instead of $this->app->bind(), which would not have applied the extra bindings. Now it does. It's backwards compatible, in that if you continue to use $this->app->bind(), it would be as functional and as broken as it was already.

In our provider, we set the default repositories, but only if they haven't already been bound. If someone has already bound it (a service provider alphabetically before Statamic), we don't want to override it. If a provider does it after Statamic, it'll just get overridden naturally.