Adjust user's avatar when augmenting ...
The avatar was not being returned in templates because $user->avatar() was the full asset url. Then, when it was being wrapped by the asset fieldtype, the raw value was the url and should have just been the path. When you try to render that in a template, the asset ends up not being found and returns null.

Since $user->avatar() is supposed to return a url, when augmenting, we'll explicitly get the Value class instead of the url. This would bypass the re-wrapping in the wrapValue method.

Moved user avatar methods into a trait just to keep things a little more organized, since more avatar related methods are added.

hasAvatarField and avatarFieldUrl are now public so they can be called from within the AugmentedUser class.

Closes #2017
Closes #2443