Deal with paths until necessary ...
There's a lot of overhead involved in converting paths to asset instances and hydrating them (reading the meta files). Especially on folders with thousands of assets. Even worse on S3 when it has to do a ton of API requests.

Use paths and natural limiting/offset unless we need asset instances, eg. when ordering or filtering with wheres.