store the paths in the cache as if it were a "path" index ...
if you were to query ->where('path', ...) a separate "path" index would be generated, but it would be identical to the _paths array. Might as well use that, and it prevents another index being created.

Especially useful for assets, where it would end up loading the meta file for no real reason. More API calls.