Support replacing the built in ProseMirror nodes and marks with custom nodes and marks
I've been looking into extending the Paragraph and Heading nodes to allow the addition of extra attributes. While it is currently possible to add new node types to the ProseMirror renderer, it is not possible to replace the built in ones. This functionality exists in the ProseMirror Renderer (Statamic is already using it to replace the Image node), but the Augmentor class does not expose a way for addons etc. to access it. This commit adds that ability.