Adjust how modified items are retrieved ...
The item being retrieved during handling of file changes should be what's on disk. It shouldn't grab it from the cache as it might be outdated.

The fresh item instance will be cached later on in handleFileChanges.

For terms, rather than mess with modifying paths and appending sites, leave it alone. Instead, let getItemFromModifiedPath return multiple items. Return all the localized terms there.