Handle creating entries and nav items ...
When you add a regular link, the page editor opens immediately and uses the 'creating' flag. At that point the page doesn't exist in the tree at all. When it opens, it'll ajax the create endpoint to get the meta etc.

When you add an entry, it doesn't pop open the editor. It adds the item to the tree and data. It just doesnt add "values", which is what we'll use to know whether to ajax load data.