Make bard hide custom buttons that have not been selected
Custom buttons that have not been selected in the blueprint editor
to be available for that field will now get hidden. However, this will
only happen if you return an array of button in the $bard.buttons callback
and not if you manipulate the buttons array directly.

The Statamic.$bard.buttons callback handler receives a new property `selectedButtons`.
This array of strings contains the selected buttons of the current field. The argument
will be undefined when being called from the blueprint editor, so that you can
distinguish whether you should add a button or not.

Example:

Statamic.$bard.buttons((buttons, selectedButtons) => {
    if (!selectedButtons || selectedButtons.includes('lead')) {
        buttons.push({
            name: 'lead',
            text: 'Lead Text',
            command: 'lead',
            icon: 'lead',
            component: LeadButton,
        });
    }
});

However, if you simply return an array of buttons bard will take care of the visible
state itself. This is meant for backwards compatibility reasons.