Makes working with fieldset prefixes a bit easier in Antlers
Provides a new handle_prefix parameter that will be applied to all variable lookups within the scope tag and the partial tag. These calls can be nested, and the first one that matches a variable is what is returned.

If no prefixed match is found, normal/default variable handling behavior is then used, and then the normal Cascade logic is used. 🎉

```

{{# Dynamically adding a trailing _ #}}
{{ scope :handle_prefix="prefix ?= prefix + '_'" }}
    {{ title }}
{{ /scope }}

{{# Hardcoded prefix in a partial tag #}}
{{ partial:partial-name handle_prefix="prefix_" }}
```