Add ArrayAccess & property access to the Value class
As discussed on Discord.

- Adds magic `__get()` method
- Adds `ArrayAccess`

This makes sure you'll never have to call `->value()` on a Value object anymore in Blade templates. This is because when, for example, you call `$page->sets` where `sets` is a replicator, only the top level will be augmented, any fields inside the sets will still be `Value` objects.

I've added an implementation for `offsetSet` and `offsetUnset`, but these probably won't and shouldn't work, it might be better to leave them empty and throw an exception if you try to set them? Let me know and I'll change it.