Ensures parameter-style modifiers work in fields like Bard (antlers: true content)
This reverts the original change to Value, and implements the "Document Transformer" strategy. This strategy works on the augmented value of the field (to account for field types like Bard/Replicator/etc.).

This strategy will attempt to reverse any HTML encoding that happened within the Antlers tags, and then parses that string instead.