Allows localization variable to be saved as false.
The localisation variable is always reset as true when updating blueprints. If you have a field that is not localizable and update the blueprint for any reason, the field is sent in the PATCH request with `localizable: false` which is then unset by this code. This means if the field has been localizable before then it can't be changed to non-localizable... which causes an issue as slugs are default localizable even when multisite is not enabled.

This code checks if multisite is enabled, and prevents the field from being unset. This fixes #7082, #4471 and continues to allow the localizable field to be unset on single sites so that the blueprints remain unaffected. On multi-sites, the blueprints will be updated the next time they are saved, and will explicitly say if fields are localizable or not by adding `localizable: false` to each one.