Fix default time problem on creating new entry
When I was creating a new entry in a collection with enabled Publish Dates, the default time was always 00:00. this happened when I make time enabled on `date` field. In these lines, you've limited the date value to 10 characters, so it only includes `YYYY-MM-DD` and misses the time. I wonder why these lines exist, when the "date" field is always there and sets its own value.