Change hydrate methods from private to protected
Having a large Statamic website, where users use CP like an admin panel, sometimes users have a lot of Global sets.

We do not want to hydrate all of these global sets to a view. Some of them are used primarily for storing data we use in the backend.

I would like to have the ability to extend Cascade so that I can exclude some globals from the cascade.