Add a token-based route to trigger a phone home
By having a way to find out if a site is online and what version of Statamic/PHP it's running, we can build a more useful account dashboard for our users that has the ability to show uptime and help see the state of their sites at a glance in a way we cannot reliably do right now with a CP-only ping.

Without knowing a site key, this route does nothing, and if you do know the key, no information is exposed, it simply triggers the same Outpost ping that signing into the control panel would.