Improve nested field ids and focus behavior of some fieldtypes
Fields inside Bard, Replicator, and other nestable fieldtypes were creating duplicate HTML IDs which caused issued with the wrong input being focused.

This PR addresses that by prefixing field IDs with `fieldPathPrefix` if set.

It also sets the `fieldId` on a number of fieldtypes that didn't have it set, which allows the fields to be focused by clicking on their respective labels.

Closes #6616.