[4.x] Fix 'Call to a member function get() on null' where Bard asset doesn't exist.
If an asset used within a Bard image node does not exist (e.g. it has been deleted from the assets panel), currently a `Call to a member function get() on null` error will be thrown when Statamic attempts to fetch the alt text.

`Optional` only provides null safety for the first method call in the chain. Using the `?->` operator instead provides the desired effect.