Save parent in revision data
We can't change the tree when saving a revision because it's not published yet.

However, we *do* need to save it, which we're doing with the whole supplements thing.

I've adjusted the code in `EntriesController@extractFromFields` since we need to push the new parent back up to the frontend, without it actually being the parent.