Fix test by moving the tag before content creation ...
When the uri is evaluated, the parser caches the tag list. Since the tag isn't available at that point, when we try to use it in the view, it's missing.

Moving the tag registration earlier makes it available when the parser does the uri. It's more reflective of a real life solution anyway. Tags would be registered early.