Fix this test ...
First of all, a slug of just a slash is a bit weird. Surprising it worked to begin with.
Second, now that when generating the uri, it might end up reading the file again. Since the collectino created a line above requires slugs by default, it took the filename which was just the id, and used it as a slug. When we did ->get('/') it would 404 because it thinks the uri is actually /1.

Easy fix is to just use an explicit not-a-slash slug.