Rather than make fields ltr in css, allow passing a direction config to text fieldtype ...
The css rules were targeting the "Sites" form.
The rules only worked when you had multisite enabled and the fields are nested in a grid. If you had a single site, the fields would be top level and wouldn't get the css applied.

Simplify it by just forcing it ltr at the field level.