Actions will show a dirty state warning ...
When the form is dirty, a warning will be shown.

The message will be generic, unless overridden on the action, like the Duplicate action is doing.

Actions can opt out of the dirty state navigate-away dialog, like the Delete action is doing. In those cases, the warning won't be shown either.