[5.x] Add negative bottom margin to Textarea's character counter (#10390)
Add negative bottom margin to Textarea's character counter

Because this element is aligned to the right, it creates a larger-than-normal gap between this field and the next one below it. This change will pull the next field up by the proper amount so it looks the same whether it's there or not.