Fix confirmation modal submitted twice
When using tabs in the confirmation modal to navigate to the cancel or
submit button pressing enter would either fire the confirm event twice
on the confirm button or fire one cancel and one confirm event on the
cancel button. This commit disables the global enter-key listener on any
button in the confirmation modal to prevent double event fireing.