Add optionItems method to improve performance ...
This method gets an array of Item instances that would be used for each option.

To make it backwards compatible, a method is added to the abstract class. It's not performant, but it works.

The BasicDictionary class overrides it to make it performant.

If someone is making a dictionary that extends the base class, they don't have to override it, but they can if they want to improve the performance.

This is all done because ->options() would get the options, then calling ->get() is happening for each option, which would end up getting all the options for each one. In the case of the file dictionary for example, this means the file is read and parsed for each line in the file.