Fix markdown and code fieldtypes ...
markRaw is necessary to prevent Vue making the codemirror instance into a proxy which breaks it.

Now that teleport is being used instead of portals, the editor isn't destroyed when you go in and out of fullscreen mode, so we can initialize codemirror once on mount. We use next tick because that's when the portal is ready.