[5.x] Fix issue with localication files named like global sets
We were encountering an issue ’’’placeholderLabel(): Argument #1 ($label) must be of type ?string, array given’’’ that appeared when there is a localization file that is named like a global set. Similar issues can appear with fieldsets. They are related to the way laravel localization works when a translation file is named like a translation key, e.g. __('search") and lang/en/search.php. In this case the whole contents of a file are returned instead of a string.

The underlying issue was already encountered and had been fixed for some parts of statamic 4, so I applied the same fix to the relevant files. https://github.com/statamic/cms/pull/9525/files