Run isNotModified check
Calling this check triggers the ETag matching, and sets the notModified response code if a match is found. 
If https://github.com/laravel/framework/blob/12.x/src/Illuminate/Http/Middleware/CheckResponseForModifications.php is in the middleware stack it would do it too, it is not in the stack by default. And checking multiple times causes no issues