Do away with the `two_factor_completed` timestamp
Laravel Fortify (and the new starter kits) don't seem to have this `_completed` column.

We already have the `two_factor_confirmed_at` timestamp, which is set when the user enables 2FA.

We probably don't need a separate timestamp when the user has saved their recovery codes.

If we do end up needing this, we can add it back in future.