[6.x] Asset Browser  (#11807)
* One asset container at a time.

* Use new dropdown components

* Prevent warnings about `actionOpened`

* wip

* Move new folder events into `sharedAssetEvents` object

* Do away with unused asset container index view

It doesn't look like we link to it anywhere and it's actually broken on v5, so let's get rid of it.

Keeping the endpoing since we *do* use it via AJAX.

* We won't be navigating between containers anymore.

* Don't keep track of container here. It never changes.

* Rename thumbnail size preference key

It should be nested under the `assets` key (like columns, sort direction, etc).

It should also be in snake case.

* Ensure toggle group always has a value.

Otherwise, you could end up with no selected mode in the asset browser.

* Ability to navigate between previous/next assets

Using "cmd + left arrow" and "cmd + right arrow" (or ctrl on Windows)

* Ensure folders can be created in Grid mode, when no existing folders exist.

* Prevent the images themselves being dragged

* Drag & drop assets and folders in grid mode

* Prevent upload overlay showing when dragging assets & folders

* Remove data attributes. I never ended up using them.

* Dry up new folder stuff.

I didn't realise there was a mixin for this when I built it. Much nicer now.

* Use the same order between two asset browser components.

* Uncomment assets css

* Drag & drop assets and folders in table mode

* Prevent moving assets & folders without permission

* Add previous/next buttons

* Only save when navigating between prev/next assets when there's dirty state