Date fieldtype: pass around dates with milliseconds
Previously, we were sending an ISO-8601 date to the frontend *without* milliseconds.

However, as soon as you change it, the @internationalized/date library would output it *with* milliseconds.

This would cause the fieldtype's validation to fail as it wasn't expecting the milliseconds.

This commit fixes it by always passing around the date with milliseconds.