Handle ArrayableLinks ...
When ResolveRedirect::item() receives an object as $redirect, it would have made it all the way to the final return statement. Then, the ArrayableLink would come back to the 404 guard in resolve(), but pass the condition because its an object and not null. Finally, the url method would be called on it which would be null.

We want a 404 though, so we'll make sure to resolve the url earlier so the guard fails as expected.