Open bard set picker using forward slash ...
Names are added to extensions otherwise you get a warning of conflicting names.

Originally I tried to call $refs.setPicker.open() from within the shortcut, but that didn't work. Even if I wrapped it in a computed. After opening the picker and making a selection, it would reference a different vue component. (I checked with $.uid). Having an openSetPicker method sorted this out. 🤷‍♂️