Fix a couple of issues with the Two Factor popover
When you click a button, we need to close the popover and open the relevant modal.

I needed to move the disable component out of the Popover slot, otherwise it would close when the popover closes.

Since the Disable component was relying on its slot to call the `confirm` method and we're not using that slot anymore, it didn't really make sense for its logic to be continue being separate so I've merged it into the main TwoFactor component.