Fix recursion issue when using the Group fieldtype inside a Replicator
Previously, the `previews` computed property was returning the field previews for the entire Replicator.

This meant that the field preview for the Group field kept chaning, causing the field previews to constantly be regenerated, causing an infinite loop.

It ended up being a simple fix - we need to append the handle of the Group field to the field path prefix when getting the field previews array and getting stuff from the publish container.