Update Arr.php
The __('Expand') is used in the collection tree as button labels for expand/collapse and in the Array field type as label title for the last option 'Whether to save the array in the expanded format. Use this if you intend to have numeric values.'

In German, there is no suitable word that is correct for both and also fits the combination collapse/expand.

Therefore, I suggest renaming the label title in the array field type slightly.