sharing/publishing ...
When you edit custom values, the option to publish becomes available and takes you to statamic.com to publish in your account.

The next time you switch to the themes tab, it'll refresh / bust the cache so you will see your newly added theme.

Added some methods to the Marketplace client that'll handle busting the cache. Realized there's two layers of caching. Just going to work around that for now. Fixing is out of scope of this PR.